Added modular __init__.py structure, ignore that files in linter.yml
