+++ .github/workflows/linter.yml
@@ -19,4 +19,4 @@ jobs:
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --show-source --max-line-length=127 --statistics
+        flake8 . --count --show-source --max-line-length=127 --per-file-ignores="__init__.py:F401" --statistics
